+++ packages/upswyng-server/src/routes/api/alert/index.ts
@@ -3,88 +3,100 @@ import Alert, {
   fullAlertToAlert,
 } from "../../../models/Alert";
 import { NextFunction, Request, Response } from "express";
-import { TAlert, TAlertFull } from "@upswyng/upswyng-types";
 
+import { ObjectId } from "bson";
 import { ParamsDictionary } from "express-serve-static-core";
+import { TAlertDocument } from "../../../models/Alert";
 import { body } from "express-validator";
 import { compose } from "compose-middleware";
-import { isAdmin } from "../../../utility/authHelpers";
+import { requireAdmin } from "../../../utility/authHelpers";
 
 interface TRequest extends ParamsDictionary {
-  includeCancelled: string | null; // boolean, default true
-  includeUnapproved: string | null; // boolean, default false
-  start: string | null; // serialized Date (optional)
-  end: string | null; // serialized Date (optional)
+  _id: string | null; // if included, will try to update an existing alert
+  category: string | null;
+  color: string; // hex string
+  detail: string | null; // markdown string
+  end: string; // serialized Date
+  icon: string; // fontawesome identifier; ex: "fas fa-snowflake"
+  isApproved: string; // boolean
+  isCancelled: string; // boolean
+  start: string; // serialized Date
+  title: string;
 }
 
 /**
- * API endpoint to get a list of Alerts.
- * Returns the full alert (TAlertFull with user who created) if the user is a signed-in
- * admin. Otherwise, returns TAlert
- * Request body may have a `start` and `end` date. If not supplied,
- * the endpoint will return the last 7 days.
- * {
- *   includeCancelled: boolean | null, // default true
- *   includeUnapproved: boolean | null, // default false
- *   start: Date,
- *   end: Date, // default now if start is defined
- * }
+ * API endpoint to create an alert. Requires admin permissions.
  *
- * The response has the shape of:
+ * Successful response (201):
+ * {}
+ *
+ * Error response (4xx | 5xx):
  * {
- *   count: number, // The total number of alerts in the time frame
- *   alerts: TAlertFull[]
+ *   message: string; // human readable error message
  * }
  */
 export const post = compose([
-  body("start")
-    .toDate()
-    .exists(),
+  body("color").exists(),
   body("end")
     .toDate()
     .exists(),
-  body("includeCancelled").toBoolean(),
-  body("includeUnapproved").toBoolean(),
-
+  body("icon").exists(),
+  body("isApproved").toBoolean(),
+  body("isCancelled").toBoolean(),
+  body("start")
+    .toDate()
+    .exists(),
+  body("title").exists(),
   async (req: Request<TRequest>, res: Response, _next: NextFunction) => {
-    const admin = isAdmin(req);
-
-    const now = new Date();
-    const start: Date | null =
-      req.body.start || new Date(now.getTime() - 7 * 24 * 60 * 60 * 1000);
-    const end: Date = req.body.end || new Date();
-
-    const includeCancelled = req.body.includeCancelled !== false ? true : false;
-    const includeUnapproved = !!req.body.includeUnapproved;
-    if (start >= end) {
-      return res.status(422).json({
-        message: `The start date, ${start}, is before the end date, ${end}`,
+    let user;
+    try {
+      user = requireAdmin(req);
+    } catch (_e) {
+      return res.status(401).json({
+        message: `You are not authorized to create alerts`,
       });
     }
+    const id = req.body._id;
+    let existingAlert;
+    if (id) {
+      try {
+        existingAlert = Alert.findById(ObjectId.createFromHexString(id));
+      } catch (e) {
+        return res
+          .status(500)
+          .json({ message: `Error finding exist alert: ${e.message}` });
+      }
+      if (!existingAlert) {
+        return res
+          .status(400)
+          .json({ message: `Could not find alert with ID: ${id}` });
+      }
+    }
+
+    const start: Date = req.body.start;
+    const end: Date = req.body.end;
 
     try {
-      const alertsDocuments = await Alert.find({
-        start: { $gte: start },
-        end: { $lte: end },
-        isCancelled: { $in: [false, includeCancelled] },
-        isApproved: { $in: [true, !includeUnapproved] },
-      });
-      let alerts: TAlert[] | TAlertFull[] = await Promise.all(
-        alertsDocuments.map(alertDocumentToFullAlert)
-      );
-
-      if (!admin) {
-        // remove sensitive fields from the response if the user is not logged
-        // in as an admin
-        alerts = (alerts as TAlertFull[]).map(fullAlertToAlert);
+      const alert: TAlertDocument = existingAlert || new Alert();
+      if (!existingAlert) {
+        alert.createdBy = user._id;
       }
-
-      return res.status(200).json({ alerts, count: alerts.length });
+      alert.category = req.body.category || null;
+      alert.color = req.body.color;
+      alert.detail = req.body.detail || null;
+      alert.end = end;
+      alert.icon = req.body.icon;
+      alert.isApproved = req.body.isApproved;
+      alert.isCancelled = req.body.isCancelled;
+      alert.lastModifiedBy = user._id;
+      alert.start = start;
+      alert.title = req.body.title;
+      await alert.save();
+      return res.status(!existingAlert ? 201 : 204).json({});
     } catch (e) {
-      console.error(`Problem while fetching alerts at /api/alert: ${e}`);
-      return res.status(500).json({
-        message: `There was an error fetching the alerts: ${e.message}`,
-      });
+      return res
+        .status(500)
+        .json({ message: `Error creating alert: ${e.message}` });
     }
   },
 ]);
